6 hw

+++ University/Teacher.cs
@@ -5,7 +5,7 @@
         public List<string> Courses { get; set; } 
         public string Degree { get; set; }
 
-        public Teacher(Person person, ulong taxId, List<string> courses, string degree) : base(person, taxId)
+        public Teacher(Person person, uint taxId, List<string> courses, string degree) : base(person, taxId)
         {
             Courses = courses;
             Degree = degree;
